feat: move Note Feature to a dedicated service/controller
