+++ apps/backend/src/features/notes/index.ts
@@ -1,0 +1,4 @@
+export { default as NotesController } from './notes.controller';
+export * from './notes.route';
+export * from './notes.schema';
+export * from './notes.service';
