+++ apps/backend/src/features/requeteEtapes/requetesEtapes.type.ts
@@ -3,7 +3,6 @@ import type {
   GetRequeteEtapesQuerySchema,
   UpdateRequeteEtapeNomSchema,
   UpdateRequeteEtapeStatutSchema,
-  updateRequeteEtapeNoteBodySchema,
 } from './requetesEtapes.schema';
 
 export type RequeteEtapeCreationDto = {
@@ -13,12 +12,3 @@ export type GetRequeteEtapesQuery = z.infer<typeof GetRequeteEtapesQuerySchema>;
 export type UpdateRequeteEtapeStatutDto = z.infer<typeof UpdateRequeteEtapeStatutSchema>;
 
 export type UpdateRequeteEtapeNomDto = z.infer<typeof UpdateRequeteEtapeNomSchema>;
-
-export type CreateRequeteEtapeNoteDto = {
-  texte: string;
-  userId: string;
-  requeteEtapeId: string;
-  fileIds: string[];
-};
-
-export type UpdateRequeteEtapeNoteDto = z.infer<typeof updateRequeteEtapeNoteBodySchema>;
